collab: Extract contributor endpoints to their own module
